Download avatars instead of relying on Big Picture cache
We were getting them from the local Big Picture cache, but that cache isn't updated unless you log in to the account using Big Picture (or the new gamepadui).

This gives us a much better implementation and experience for the avatars, opening up the possiblilty for it to be opt-out instead of opt-in.

We could cache the images in the future, but this is a lot simpler to implement right now.